gramr {gramr}	R Documentation

A R package To Check for Grammar within RMarkdown documents.

Description

Provides additional checks in grammar not currently caught by the native spell check within RStudio.

[Package gramr version 0.0.0.9000 Index]