run_grammar_checker {gramr}	R Documentation

Start grammar checker shiny application Start the grammar checker. This allows the user to step through lines of text which may have grammar errors, and then write the resulting file.

Description

Start grammar checker shiny application Start the grammar checker. This allows the user to step through lines of text which may have grammar errors, and then write the resulting file.

Usage

run_grammar_checker(path)

Arguments

path
the intended filepath

Value

a shiny app is launched

Examples

# don't run during tests
# gramr::run_grammar_checker("example.rmd")

[Package gramr version 0.0.0.9000 Index]